feat: complete error, loading, success models and opertions completed in signup
