+++ milestone-master/package.json
@@ -20,13 +20,14 @@
     "react-redux": "^8.1.3",
     "react-router-dom": "^6.25.1",
     "react-scripts": "5.0.1",
+    "redux-logger": "^3.0.6",
     "redux-thunk": "^3.1.0",
     "uuid": "^10.0.0",
     "web-vitals": "^2.1.4",
     "yup": "^1.4.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "start": "set PORT=3001 && react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
     "eject": "react-scripts eject"
