+++ milestone-master/src/App.js
@@ -1,8 +1,8 @@
 import React from "react";
 import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Homepage from "./components/home/HomePage";
-import Signup from "./components/signup/Signup"; // Create this component
-import Login from "./components/login/Login"; // Create this component
+import Signup from "./components/home/Signup"; // Create this component
+import Login from "./components/home/Login"; // Create this component
 import { Box } from "@mui/material";
 import ReferenceBooks from "./components/home/ReferenceBooks";
 import styles from "./components/home/css/homepage.module.css";
