+++ milestone-master/src/components/home/slices/authSlice.js
@@ -1,17 +1,33 @@
 import { createSlice, createAsyncThunk } from '@reduxjs/toolkit';
+import axios from 'axios';
 
 // Utility function to handle fetch requests
-const handleFetch = async (url, options) => {
-    const response = await fetch(url, options);
-    if (!response.ok) {
-        const errorData = await response.json();
-        throw new Error(errorData.message || 'Something went wrong');
+const handleFetch = async (url, options = {}) => {
+    const completeUrl = `http://127.0.0.1:8000${url}`;
+    try {
+        const response = await axios({
+            url: completeUrl,
+            method: options.method || 'GET',
+            headers: options.headers || {},
+            data: options.body || {},
+        });
+        return response.data;
+    } catch (error) {
+        // Extract the error message from the response if available
+        console.log("error",error.response?.data)
+        const errors = error.response.data
+        const keys = Object.values(errors)
+        console.log("keys", keys)
+        let finalErrors = keys.flat()
+       console.log("finalErrors", finalErrors)
+
+        const errorMessage = finalErrors.join("\n");
+        throw new Error(errorMessage);
     }
-    return response.json();
 };
 
 export const signup = createAsyncThunk('auth/signup', async (userData) => {
-    const response = await handleFetch('/api/signup', {
+    const response = await handleFetch('/api/users/', {
         method: 'POST',
         headers: {
             'Content-Type': 'application/json',
@@ -41,11 +57,17 @@ export const logout = createAsyncThunk('auth/logout', async () => {
 const authSlice = createSlice({
     name: 'auth',
     initialState: { user: null, status: 'idle', error: null },
-    reducers: {},
+    reducers: {
+        resetStatus: (state) => {
+            state.status = 'idle';
+            state.error = null;
+          }
+    },
     extraReducers: (builder) => {
         builder
             .addCase(signup.fulfilled, (state, action) => {
                 state.user = action.payload;
+                state.status = "success";
             })
             .addCase(login.fulfilled, (state, action) => {
                 state.user = action.payload;
@@ -54,12 +76,17 @@ const authSlice = createSlice({
                 state.user = null;
             })
             .addCase(signup.rejected, (state, action) => {
-                state.error = action.error.message;
+                state.error = action.error;
+                state.status = "failed";
             })
             .addCase(login.rejected, (state, action) => {
                 state.error = action.error.message;
+            })
+            .addCase(signup.pending, (state, action) => {
+                state.status = "pending";
             });
     },
 });
 
+export const { resetStatus } = authSlice.actions;
 export default authSlice.reducer;
